requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
gunicorn==25.0.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
zstandard==0.25.0
